app_default_25_11_27_03_41_13/app_default_25_11_27_03_41_13/requirements.txt
fastapi==0.111.3
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.34
psycopg[binary]==3.1.2
python-jose==4.4.0
passlib[bcrypt]==1.7.4
pytest==7.4.2
httpx==0.24.1
python-dotenv==1.0.0
